+++ utility/types/tx_result.go
@@ -9,21 +9,21 @@ import (
 
 // INVESTIGATE: Look into a way of removing this type altogether or from shared interfaces.
 
-var _ modules.TxResult = &DefaultTxResult{}
+var _ modules.TxResult = &TxResult{}
 
-func (txr *DefaultTxResult) Bytes() ([]byte, error) {
+func (txr *TxResult) Bytes() ([]byte, error) {
 	return codec.GetCodec().Marshal(txr)
 }
 
-func (*DefaultTxResult) FromBytes(bz []byte) (modules.TxResult, error) {
-	result := new(DefaultTxResult)
+func (*TxResult) FromBytes(bz []byte) (modules.TxResult, error) {
+	result := new(TxResult)
 	if err := codec.GetCodec().Unmarshal(bz, result); err != nil {
 		return nil, err
 	}
 	return result, nil
 }
 
-func (txr *DefaultTxResult) Hash() ([]byte, error) {
+func (txr *TxResult) Hash() ([]byte, error) {
 	bz, err := txr.Bytes()
 	if err != nil {
 		return nil, err
@@ -31,17 +31,17 @@ func (txr *DefaultTxResult) Hash() ([]byte, error) {
 	return txr.HashFromBytes(bz)
 }
 
-func (txr *DefaultTxResult) HashFromBytes(bz []byte) ([]byte, error) {
+func (txr *TxResult) HashFromBytes(bz []byte) ([]byte, error) {
 	return crypto.SHA3Hash(bz), nil
 }
 
-func TxToTxResult(tx *coreTypes.Transaction, height int64, index int, signer, recipient, msgType string, err Error) (*DefaultTxResult, Error) {
+func TxToTxResult(tx *coreTypes.Transaction, height int64, index int, signer, recipient, msgType string, err Error) (*TxResult, Error) {
 	txBytes, er := tx.Bytes()
 	if er != nil {
 		return nil, ErrProtoMarshal(er)
 	}
 	code, errString := int32(0), ""
-	return &DefaultTxResult{
+	return &TxResult{
 		Tx:            txBytes,
 		Height:        height,
 		Index:         int32(index),